+++ ichnaea/data/tests/test_ocid.py
@@ -2,7 +2,7 @@ import csv
 import os
 import re
 from contextlib import contextmanager
-from datetime import datetime
+from datetime import datetime, timedelta
 
 import boto
 from mock import MagicMock, patch
@@ -59,6 +59,9 @@ class FakeTask(object):
 class TestExport(object):
 
     def test_local_export(self, celery, session):
+        now = util.utcnow()
+        today = now.date()
+        long_ago = now - timedelta(days=367)
         cell_fixture_fields = (
             'radio', 'cid', 'lat', 'lon', 'mnc', 'mcc', 'lac')
         base_cell = CellShardFactory.build(radio=Radio.wcdma)
@@ -81,11 +84,14 @@ class TestExport(object):
 
         # add one incomplete / unprocessed cell
         CellShardFactory(cid=210, lat=None, lon=None, **cell_key)
+        # add one really old cell
+        CellShardFactory(cid=220, created=long_ago, modified=long_ago,
+                         last_seen=long_ago.date(), **cell_key)
         session.commit()
 
         with util.selfdestruct_tempdir() as temp_dir:
             path = os.path.join(temp_dir, 'export.csv.gz')
-            write_stations_to_csv(session, path)
+            write_stations_to_csv(session, path, today)
 
             with util.gzip_open(path, 'r') as gzip_wrapper:
                 with gzip_wrapper as gzip_file:
@@ -118,7 +124,12 @@ class TestExport(object):
             assert pattern.search(method.call_args[0][0])
 
     def test_export_full(self, celery, session):
+        now = util.utcnow()
+        long_ago = now - timedelta(days=367)
         CellShardFactory.create_batch(10, radio=Radio.gsm)
+        CellShardFactory(
+            radio=Radio.gsm, created=long_ago,
+            modified=long_ago, last_seen=long_ago.date())
         session.commit()
         pattern = re.compile(
             r'MLS-full-cell-export-\d+-\d+-\d+T000000\.csv\.gz')