+++ CHANGES.rst
@@ -14,6 +14,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- #490: Limit full cell export to cells which were modified in the last year.
+
 - Use the accuracy_radius field from the GeoIP database.
 
 - Remove ipf fallback from responses for queries based only on GeoIP data.